+++ test/unit/slurm/test_slurm_config_generator.py
@@ -128,6 +128,13 @@ def test_generate_slurm_config_files_memory_scheduling(
             ],
             id="Case with Slurm Accounting passing DatabaseName",
         ),
+        pytest.param(
+            "sample_input_external_slurmdbd.yaml",
+            [
+                "slurm_parallelcluster_external_slurmdbd.conf",
+            ],
+            id="Case with Slurmdbd daemon external to the cluster",
+        ),
     ],
 )
 def test_generate_slurm_config_files_slurm_accounting(mocker, test_datadir, tmpdir, input_config, expected_outputs):